requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
argcomplete==1.10.3
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.8.2
blinker==1.7.0
blis==0.7.11
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
compressed_rtf==1.0.6
confection==0.1.4
crashtest==0.4.1
cryptography==41.0.7
cymem==2.0.8
distlib==0.3.8
Django==5.0.2
django-ckeditor==6.7.0
django-js-asset==2.2.0
docx2txt==0.8
dulwich==0.21.7
ebcdic==1.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
extract-msg==0.28.7
fastapi==0.108.0
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.1
Flask-Cors==4.0.0
gitdb==4.0.11
GitPython==3.1.40
gym==0.25.2
gym-notices==0.0.8
h11==0.14.0
hashids==1.3.1
idna==3.6
IMAPClient==2.1.0
importlib-metadata==6.11.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jericho==3.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonlines==4.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keyring==24.3.0
langcodes==3.3.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mementos==1.3.1
more-itertools==10.1.0
msgpack==1.0.7
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
olefile==0.47
packaging==23.2
pafy==0.5.5
pandas==2.2.0
pdfminer.six==20191110
pdfminer3==2018.12.3.0
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==3.11.0
plotly==5.18.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.1
ptyprocess==0.7.0
pyarrow==14.0.2
pycparser==2.21
pycryptodome==3.19.1
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyPDF2==3.0.1
pyproject_hooks==1.0.0
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.8.2
python-pptx==0.6.23
pytz==2023.3.post1
rapidfuzz==3.6.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.16.2
scikit-learn==1.4.0
scipy==1.12.0
setuptools==69.0.3
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
sqlparse==0.4.4
srsly==2.4.8
starlette==0.32.0.post1
streamlit==1.29.0
TatSu==4.4.0
tenacity==8.2.3
textract==1.6.5
textworld @ https://github.com/MarcCote/TextWorld/archive/refs/heads/inform_1.10.zip#sha256=3048f3b44e6a9f906b509160654e668402f76ad17da81851a741ad7b2eaf0f41
thinc==8.2.2
threadpoolctl==3.2.0
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
trove-classifiers==2024.1.31
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
validators==0.22.0
virtualenv==20.25.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
wheel @ file:///Users/divyanshgupta/Desktop/projects/Smart_Resume_Analyser_App/wheel-0.42.0-py3-none-any.whl#sha256=177f9c9b0d45c47873b619f5b650346d632cdc35fb5e4d25058e09c9e581433d
xattr==0.10.1
xlrd==1.2.0
XlsxWriter==3.1.9
zipp==3.17.0
